requirements.txt
Django==3.2.16
djangorestframework==3.15.1
drf_yasg==1.21.7
django-cors-headers==4.3.1